Refactor and test getTagMetadata + remove dummy data from resource
